events: add Kafka settings
