+++ django/src/app/conf/events.py
@@ -1,0 +1,15 @@
+KAFKA_HOST = 'localhost:9092'
+
+
+class Topics:
+    """
+    list of available topics to produce and consume
+    """
+
+    # tasks: CUD events
+    tasks_stream = 'tasks-stream'
+    # tasks: business events
+    tasks = 'tasks'
+
+    # users: CUD events
+    users_stream = 'users-stream'
