+++ django/src/app/settings.py
@@ -18,6 +18,7 @@ include(
     "conf/auth.py",
     "conf/boilerplate.py",
     "conf/db.py",
+    "conf/events.py",
     "conf/healthchecks.py",
     "conf/http.py",
     "conf/i18n.py",
