+++ django/src/tasks/api/views/tasks.py
@@ -5,12 +5,13 @@ from rest_framework import permissions, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 
+from app.settings import Topics
+
 from tasks.models import Task, TaskCost, TaskStatus, TaskUser
 from tasks.api.serializers import TaskSerializer, TaskUserSerializer
 from tasks.producer import Producer
 
 from tasks.tests.mock import produce  # FIXME change for real kafka producer
-
 
 
 # FIXME users can't be created here, only replicated via kafka
@@ -46,7 +47,7 @@ class TaskViewSet(viewsets.ModelViewSet):
 
         serializer.save(user=user)
         # CUD event: task created
-        p.produce(topic='tasks-stream', key='task-created', value=json.dumps(serializer.data))
+        p.produce(topic=Topics.tasks_stream, key='task-created', value=json.dumps(serializer.data))
 
     def perform_update(self, serializer):
         initial_status = serializer.instance.status
@@ -55,7 +56,7 @@ class TaskViewSet(viewsets.ModelViewSet):
         serializer.save(**serializer.validated_data)
         if initial_status != new_status:
             # business event: status changed
-            p.produce(topic='tasks', key='task-status-updated', value=json.dumps(serializer.data))
+            p.produce(topic=Topics.tasks, key='task-status-updated', value=json.dumps(serializer.data))
 
     # FIXME change permissions to IsAdmin
     @action(detail=False, methods=['post'], url_path='reassign', url_name='reassign', permission_classes=[permissions.AllowAny])
